fix to organize files
