+++ lib/ui/layout/default_layout.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:smart_order_app/ui/component/app_bar/simple_app_bar.dart';
 import 'package:smart_order_app/ui/component/app_bar/simple_bottom_app_bar.dart';
-import 'package:smart_order_app/ui/component/simple_drawer.dart';
+import 'package:smart_order_app/ui/component/drawer/simple_drawer.dart';
 
 class BodyOption {
   final bool dense;
