+++ lib/ui/page/management/reason/list/page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:smart_order_app/ui/component/delete_confirmation_dialog.dart';
+import 'package:smart_order_app/ui/component/dialog/delete_confirmation_dialog.dart';
 import 'package:smart_order_app/ui/component/error_message.dart';
 import 'package:smart_order_app/ui/component/loader.dart';
 import 'package:smart_order_app/ui/error_handler_mixin.dart';
