+++ lib/ui/page/management/scene/edit/page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:smart_order_app/domain/entity/scene.dart';
-import 'package:smart_order_app/ui/component/form_error_message.dart';
+import 'package:smart_order_app/ui/component/form/form_error_message.dart';
 import 'package:smart_order_app/ui/error_handler_mixin.dart';
 import 'package:smart_order_app/ui/layout/default_layout.dart';
 import 'package:smart_order_app/ui/page/management/scene/edit/form/scene_edit_form_controller.dart';
