+++ lib/ui/page/order/select/page.dart
@@ -4,9 +4,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:smart_order_app/domain/valueObject/id.dart';
 import 'package:smart_order_app/ui/component/error_message.dart';
 import 'package:smart_order_app/ui/component/loader.dart';
-import 'package:smart_order_app/ui/component/navigation_button.dart';
-import 'package:smart_order_app/ui/component/simple_checkbox_list_tile.dart';
-import 'package:smart_order_app/ui/component/typography.dart';
+import 'package:smart_order_app/ui/component/button/navigation_button.dart';
+import 'package:smart_order_app/ui/component/form/simple_checkbox_list_tile.dart';
+import 'package:smart_order_app/ui/component/typography/section_title.dart';
 import 'package:smart_order_app/ui/form/form_creation_status.dart';
 import 'package:smart_order_app/ui/layout/default_layout.dart';
 import 'package:smart_order_app/ui/page/order/display/page.dart';
